backend/requirements.txt
fastapi==0.121.0
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
pandas==2.3.3
python-dotenv==1.2.1